[Chore] new file `defaultConfiguration.ts`
